Progress refactoring app
