+++ packages/mechanic/app/App.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { Switch, Route, withRouter } from "react-router-dom";
 
-import { Function } from "./pages/Function";
+import { Controls } from "./pages/Controls";
 import { NotFound } from "./pages/NotFound";
 import { Nav } from "./pages/Nav";
 
@@ -20,9 +20,9 @@ const AppComponent = () => {
             key={`route-${name}`}
             path={i == 0 ? ["/", `/${name}`] : `/${name}`}
             render={() => (
-              <Function name={name} exports={functions[name]}>
+              <Controls name={name} exports={functions[name]}>
                 <Nav name={name} functionsNames={functionNames} />
-              </Function>
+              </Controls>
             )}
           />
         ))}
