+++ packages/mechanic/app/pages/utils/useShortcuts.js
@@ -1,0 +1,13 @@
+import { useEffect } from "react";
+import Mousetrap from "mousetrap";
+
+const useShortcuts = handleExport => {
+  useEffect(() => {
+    Mousetrap.bind("command+e", handleExport);
+    return () => {
+      Mousetrap.unbind("command+e");
+    };
+  });
+};
+
+export { useShortcuts };
