+++ packages/mechanic/src/function-set-up.js
@@ -7,12 +7,16 @@ import { MechanicError } from "./mechanic-error";
  * design functions export objects.
  */
 const setUp = functions => {
-  console.info("Setting up design functions!");
   const engines = {};
   Object.keys(functions).forEach(
     functionName => (engines[functionName] = functions[functionName].settings.engine)
   );
   let curEngine = null;
+  window.run = () => {
+    alert(
+      "The engine and/or function may still be loading, try again in a bit. If this keeps happening, post an issue!"
+    );
+  };
   window.initEngine = functionName => {
     if (engines[functionName] === undefined) {
       window.run = () => {
@@ -23,7 +27,6 @@ const setUp = functions => {
       throw new MechanicError(`No defined engine for: ${functionName}`);
     } else if (engines[functionName] !== curEngine) {
       console.info("Setting mechanic engine for:", functionName);
-      // TODO: Kill existing sketch if running?
       curEngine = engines[functionName];
       window.run = (functionName, values, isPreview) => {
         // TODO: Do performance stats here?
@@ -33,6 +36,7 @@ const setUp = functions => {
       };
     }
   };
+  console.info("Design function definitions set.");
 };
 
 export { setUp };
