+++ packages/mechanic/src/mechanic.js
@@ -67,12 +67,10 @@ export class Mechanic {
     if (!this.exportInit) {
       this.exportInit = true;
       this.serializer = new XMLSerializer();
-      if (this.settings.animated) {
-        this.videoWriter = new WebMWriter({
-          quality: 0.95,
-          frameRate: 60
-        });
-      }
+      this.videoWriter = new WebMWriter({
+        quality: 0.95,
+        frameRate: 60
+      });
     }
 
     if (validation.isSVG(el)) {
@@ -98,10 +96,7 @@ export class Mechanic {
   async done(el) {
     if (!this.settings.animated) {
       if (validation.isSVG(el)) {
-        // This conditional is a patch to an error, needs to be revised:
-        if (!this.serializer) {
-          this.serializer = new XMLSerializer();
-        }
+        this.serializer = new XMLSerializer();
         this.svgData = svgToDataUrl(el, this.serializer);
       } else {
         this.canvasData = el.toDataURL();
@@ -111,6 +106,7 @@ export class Mechanic {
         // This is slow. We should figure out a way to draw into canvas on every frame
         // or at least do Promise.all
         const cacheCanvas = document.createElement("canvas");
+        cacheCanvas.setAttribute("id", "mechanic-cache-canvas");
         cacheCanvas.width = this.svgSize.width;
         cacheCanvas.height = this.svgSize.height;
         for (let i = 0; i < this.svgFrames.length; i++) {
@@ -119,6 +115,10 @@ export class Mechanic {
         }
       }
       this.videoData = await this.videoWriter.complete();
+      if (validation.isSVG(el)) {
+        const cacheCanvas = document.getElementById("mechanic-cache-canvas");
+        cacheCanvas.remove();
+      }
     }
     this.isDone = true;
   }
